backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.34
python-dotenv==1.0.1
openai==1.107.2
orjson==3.10.7
pytest==8.3.2
requests==2.32.3 